Split applicaiton-actions into two files
