+++ src/frontend/app/shared/components/list/list-types/app-sevice-bindings/app-service-binding-card/app-service-binding-card.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material';
 import { Store } from '@ngrx/store';
-import { BehaviorSubject } from 'rxjs';
+import { BehaviorSubject } from 'rxjs/BehaviorSubject';
 import { Observable } from 'rxjs/Observable';
 import { map, tap, withLatestFrom } from 'rxjs/operators';
 import { Subscription } from 'rxjs/Subscription';
@@ -9,7 +9,7 @@ import { Subscription } from 'rxjs/Subscription';
 import { IService, IServiceBinding, IServiceInstance } from '../../../../../../core/cf-api-svc.types';
 import { EntityServiceFactory } from '../../../../../../core/entity-service-factory.service';
 import { ApplicationService } from '../../../../../../features/applications/application.service';
-import { DeleteAppServiceBinding } from '../../../../../../store/actions/application.actions';
+import { DeleteAppServiceBinding } from '../../../../../../store/actions/application-service-routes.actions';
 import { GetServiceInstance } from '../../../../../../store/actions/service-instances.actions';
 import { GetService } from '../../../../../../store/actions/service.actions';
 import { AppState } from '../../../../../../store/app-state';
@@ -17,9 +17,9 @@ import { entityFactory, serviceInstancesSchemaKey, serviceSchemaKey } from '../.
 import { APIResource, EntityInfo } from '../../../../../../store/types/api.types';
 import { AppEnvVarsState } from '../../../../../../store/types/app-metadata.types';
 import { AppChip } from '../../../../chips/chips.component';
+import { EnvVarViewComponent } from '../../../../env-var-view/env-var-view.component';
 import { MetaCardMenuItem } from '../../../list-cards/meta-card/meta-card-base/meta-card.component';
 import { TableCellCustom } from '../../../list-table/table-cell/table-cell-custom';
-import { EnvVarViewComponent } from '../../../../env-var-view/env-var-view.component';
 
 @Component({
   selector: 'app-app-service-binding-card',
@@ -69,7 +69,6 @@ export class AppServiceBindingCardComponent extends TableCellCustom<APIResource<
           new GetService(o.entity.entity.service_guid, this.appService.cfGuid),
           true
         ).entityObs$;
-
         const tags = [];
         o.entity.entity.tags.forEach(t => {
           tags.push({
