+++ src/frontend/app/store/reducers/application-route.reducer.ts
@@ -1,4 +1,4 @@
-import { ASSIGN_ROUTE_SUCCESS, AssociateRouteWithAppApplication } from '../actions/application.actions';
+import { ASSIGN_ROUTE_SUCCESS, AssociateRouteWithAppApplication } from '../actions/application-service-routes.actions';
 import { BaseRouteAction, RouteEvents } from '../actions/route.actions';
 import { APIResource } from '../types/api.types';
 import { APISuccessOrFailedAction } from '../types/request.types';
